refactor: DistanceSummary 로직 분리
