+++ src/axois/index.ts
@@ -1,7 +1,9 @@
 import axios from 'axios';
 
+export const baseURL = process.env.NEXT_PUBLIC_API_URL;
+
 export const api = axios.create({
-  baseURL: process.env.NEXT_PUBLIC_API_URL,
+  baseURL,
   timeout: 10000,
   headers: {
     'Content-Type': 'application/json',
